Fix int truncation of 64-bit usmap enum values

+++ CUE4Parse/MappingsProvider/Usmap/UsmapParser.cs
@@ -97,7 +97,7 @@ public class UsmapParser
                 {
                     var value = Ar.Read<ulong>();
                     var name = Ar.ReadName(nameLut)!;
-                    enumNames[(int)value] = name;
+                    enumNames[(long)value] = name;
                 }
             }
             else
